Dragonfly: Optimisations and improvements
Initialise the properties with the right value directly on creation
instead of setting the value afterwards
Remove calls to get the values from the device when it's not required

Remove some warnings

Remove further logging now useless

+++ DeviceAdapters/Dragonfly/TIRFModeFloatSubProperty.cpp
@@ -66,7 +66,7 @@ int CTIRFModeFloatSubProperty::OnChange( MM::PropertyBase * Prop, MM::ActionType
     {
       if ( vRequestedValue >= (double)vMin && vRequestedValue <= (double)vMax )
       {
-        if ( !DeviceWrapper_->Set( vRequestedValue ) )
+        if ( !DeviceWrapper_->Set( (float)vRequestedValue ) )
         {
           float vDeviceValue;
           DeviceWrapper_->Get( &vDeviceValue );
